--- lint-staged.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  "**/*.ts?(x)": (filenames) => [
-    `eslint ${filenames.join(" ")} --quiet --fix`,
-    `prettier --write ${filenames.join(" ")}`,
-  ],
-};
