--- sandpack-react/.env.example
@@ -1,1 +1,0 @@
-CODESANDBOX_ENV=development
