+++ sandpack-react/src/components/CodeViewer/index.tsx
@@ -2,8 +2,8 @@ import { useClasser } from "@code-hike/classer";
 import * as React from "react";
 
 import type { SandpackInitMode } from "../..";
-import { RunButton } from "../../common/RunButton";
-import { SandpackStack } from "../../common/Stack";
+import { RunButton } from "../common/RunButton";
+import { SandpackStack } from "../common/Stack";
 import { useActiveCode } from "../../hooks/useActiveCode";
 import { useSandpack } from "../../hooks/useSandpack";
 import { THEME_PREFIX } from "../../styles";
