+++ sandpack-react/src/components/FileTabs/index.tsx
@@ -2,7 +2,7 @@ import { useClasser } from "@code-hike/classer";
 import * as React from "react";
 
 import { useSandpack } from "../../hooks/useSandpack";
-import { CloseIcon } from "../../icons";
+import { CloseIcon } from "../icons";
 import { css, THEME_PREFIX } from "../../styles";
 import { buttonClassName } from "../../styles/shared";
 import { classNames } from "../../utils/classNames";
