+++ sandpack-react/src/components/Preview/index.tsx
@@ -5,10 +5,10 @@ import type {
 } from "@codesandbox/sandpack-client";
 import * as React from "react";
 
-import { ErrorOverlay } from "../../common/ErrorOverlay";
-import { LoadingOverlay } from "../../common/LoadingOverlay";
-import { OpenInCodeSandboxButton } from "../../common/OpenInCodeSandboxButton";
-import { SandpackStack } from "../../common/Stack";
+import { ErrorOverlay } from "../common/ErrorOverlay";
+import { LoadingOverlay } from "../common/LoadingOverlay";
+import { OpenInCodeSandboxButton } from "../common/OpenInCodeSandboxButton";
+import { SandpackStack } from "../common/Stack";
 import { useSandpackClient } from "../../hooks";
 import { css, THEME_PREFIX } from "../../styles";
 import { classNames } from "../../utils/classNames";
