+++ sandpack-react/src/components/Tests/SandpackTests.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
-import { SandpackStack } from "../../common";
-import { Loading } from "../../common/Loading";
+import { SandpackStack } from "../common";
+import { Loading } from "../common/Loading";
 import { useSandpackTheme, useSandpackClient } from "../../hooks";
 import { css, THEME_PREFIX } from "../../styles";
 import { classNames } from "../../utils/classNames";
