+++ sandpack-react/src/components/common/ErrorOverlay.tsx
@@ -1,14 +1,14 @@
 import { useClasser } from "@code-hike/classer";
 import * as React from "react";
 
-import { useErrorMessage } from "../hooks/useErrorMessage";
-import { THEME_PREFIX } from "../styles";
+import { useErrorMessage } from "../../hooks/useErrorMessage";
+import { THEME_PREFIX } from "../../styles";
 import {
   absoluteClassName,
   errorClassName,
   errorMessageClassName,
-} from "../styles/shared";
-import { classNames } from "../utils/classNames";
+} from "../../styles/shared";
+import { classNames } from "../../utils/classNames";
 
 /**
  * @category Components
