+++ sandpack-react/src/components/common/RunButton.tsx
@@ -1,11 +1,11 @@
 import { useClasser } from "@code-hike/classer";
 import * as React from "react";
 
-import { useSandpack } from "../hooks/useSandpack";
+import { useSandpack } from "../../hooks/useSandpack";
 import { RunIcon } from "../icons";
-import { css, THEME_PREFIX } from "../styles";
-import { roundedButtonClassName, buttonClassName } from "../styles/shared";
-import { classNames } from "../utils/classNames";
+import { css, THEME_PREFIX } from "../../styles";
+import { roundedButtonClassName, buttonClassName } from "../../styles/shared";
+import { classNames } from "../../utils/classNames";
 
 const runButtonClassName = css({
   position: "absolute",
