+++ sandpack-react/src/styles/themeContext.tsx
@@ -1,13 +1,8 @@
 import { useClasser } from "@code-hike/classer";
 import * as React from "react";
 
-import {
-  createTheme,
-  css,
-  THEME_PREFIX,
-  standardizeStitchesTheme,
-} from "../styles";
-import { standardizeTheme } from "../styles";
+import { createTheme, css, THEME_PREFIX, standardizeStitchesTheme } from ".";
+import { standardizeTheme } from ".";
 import { defaultLight } from "../themes";
 import type { SandpackTheme, SandpackThemeProp } from "../types";
 import { classNames } from "../utils/classNames";
